requirement.txt
amqp==2.6.1
apipkg==1.5
appdirs==1.4.4
appnope==0.1.2
arabic-reshaper==2.1.3
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.3.0
black==21.9b0
bleach==3.2.2
blinker==1.4
boto3==1.17.61
botocore==1.20.61
Brlapi==0.8.2
cairocffi==1.3.0
CairoSVG==2.5.2
celery==4.4.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
command-not-found==0.3
coverage==5.3.1
cryptography==3.4.7
cssselect2==0.4.1
cupshelpers==1.0
dateutils==0.6.12
dbus-python==1.2.16
debugpy==1.5.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
Django==3.1.3
django-bulk-sync==3.2.0
django-celery-beat==2.2.0
django-codemirror2==0.2
django-cors-headers==3.6.0
django-cte==1.1.4
django-db-connection-pool==1.0.6
django-db-constraints==0.3.0
django-debug-toolbar==3.2.4
django-extensions==3.0.9
django-filter==2.4.0
django-fsm==2.7.1
django-js-asset==1.2.2
django-mptt==0.9.1
django-rest==0.8.7
django-rest-framework-mongoengine==3.4.1
django-river==3.2.2
django-searchable-encrypted-fields==0.1.9
django-simple-history==2.12.0
django-storages==1.11.1
django-timezone-field==4.1.2
djangorestframework==3.12.2
dnspython==1.16.0
docopt==0.6.2
entrypoints==0.3
eventlet==0.31.1
execnet==1.8.0
filelock==3.0.12
flower==0.9.7
future==0.18.2
fuzzywuzzy==0.18.0
gevent==21.1.2
graphviz==0.15
greenlet==1.0.0
gunicorn==20.0.4
gyp==0.1
h11==0.8.1
h2==3.2.0
hidpidaemon==18.4.6
hpack==3.0.0
html5lib==1.1
http3==0.6.7
httpie==2.4.0
httplib2==0.18.1
humanize==3.5.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
jeepney==0.7.1
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kernelstub==3.1.4
keyring==23.0.1
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
louis==3.18.0
macaroonbakery==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mock==4.0.3
mongoengine==0.23.1
more-itertools==4.2.0
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
netifaces==0.10.9
newrelic==7.4.0.172
notebook==6.2.0
num2words==0.5.10
numpy==1.19.4
oauthlib==3.1.0
packaging==20.8
pandas==1.1.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.1
pathspec==0.9.0
paytmchecksum==1.7.0
pdfkit==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipenv==11.9.0
pisa==3.0.33
platformdirs==2.4.0
pluggy==0.13.1
pokemon-terminal @ git+https://github.com/LazoCoder/Pokemon-Terminal.git@ed8f7f4ebab6794e96398865f0f69df8f25048d0
pop-transition==1.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.26
protobuf==3.12.4
psutil==5.8.0
psycopg2-binary==2.8.6
ptpython==3.0.19
ptyprocess==0.7.0
py==1.10.0
pycairo==1.16.2
pycparser==2.20
pycryptodome==3.9.9
pycups==2.0.1
pydbus==0.6.0
pydot==1.4.1
pydotplus==2.0.2
pyenchant==3.2.1
pygame==2.0.3
Pygments==2.7.4
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
pyRFC3339==1.1
pyrsistent==0.17.3
pytest==6.2.1
pytest-django==4.1.0
pytest-forked==1.3.0
pytest-xdist==2.2.0
python-apt===2.2.1pop0-1628521831-21.10-5dbe6cf
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
python-Levenshtein==0.12.2
python-magic==0.4.24
python-xlib==0.29
pytz==2020.4
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.3.1
pyzmq==21.0.1
qtconsole==5.0.2
QtPy==1.9.0
redis==3.5.3
repolib==1.5.2
repoman==1.4.0
reportlab==3.5.67
requests==2.25.0
requests-async==0.6.2
requests-toolbelt==0.9.1
rfc3986==1.4.0
river-admin==0.6.0
s3transfer==0.4.2
scipy==1.6.2
scp==0.13.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
Send2Trash==1.5.0
sessioninstaller==0.0.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
systemd-python==234
tblib==1.7.0
terminado==0.9.2
testpath==0.4.4
tinycss2==1.1.1
toml==0.10.2
tomli==1.2.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unrpa==2.3.0
urllib3==1.26.2
vine==1.3.0
virtualenv==20.4.6+ds
virtualenv-clone==0.3.0
wadllib==1.3.5
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xdg==5
xhtml2pdf==0.2.5
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==1.4.0
xmltodict==0.12.0
yapf==0.31.0
youtube-dl==2021.6.6
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.3.0
